requirements/base.txt
Django==1.9.8
mysql-python==1.2.5
mysqlclient==1.3.7
djangorestframework==3.4.3